ToString method added to make it easy to work with employe information.
The ToString() method is overridden in the Employee class to provide a custom string representation of an object of that class. In C#, when you call ToString() on an object, it returns a string that represents the object. By default, the ToString() method in the Object class (the base class for all classes in C#) returns the fully qualified name of the type.